F_T30603 move MapBundle repositories to core
- adjust use statements
