+++ demosplan/DemosPlanCoreBundle/Logic/Map/MapService.php
@@ -24,9 +24,9 @@ use demosplan\DemosPlanCoreBundle\Logic\EntityHelper;
 use demosplan\DemosPlanCoreBundle\Logic\FileService;
 use demosplan\DemosPlanCoreBundle\Logic\HttpCall;
 use demosplan\DemosPlanCoreBundle\Logic\Procedure\MasterTemplateService;
+use demosplan\DemosPlanCoreBundle\Repository\GisLayerCategoryRepository;
+use demosplan\DemosPlanCoreBundle\Repository\MapRepository;
 use demosplan\DemosPlanCoreBundle\Utilities\DemosPlanTools;
-use demosplan\DemosPlanMapBundle\Repository\GisLayerCategoryRepository;
-use demosplan\DemosPlanMapBundle\Repository\MapRepository;
 use demosplan\DemosPlanMapBundle\Utilities\MapScreenshotter;
 use demosplan\DemosPlanMapBundle\ValueObject\MapOptions;
 use demosplan\DemosPlanProcedureBundle\Logic\ProcedureService;
